+++ client/test/scripts/db/gamelist.test.js
@@ -5,7 +5,7 @@ import * as gamelist from '../../../scripts/db/gamelist';
 
 test('gamelist, should getAllNames', (t) => {
   const result = gamelist.getAllNames();
-  t.is(result.length, 5);
+  t.is(result.length, 6);
 });
 
 test('gamelist, should getByName', (t) => {